refactor(attention): simplify sparse MLA helpers
Rename the shared FlashInfer sparse-MLA helper module and structure fallback availability around explicit unsupported cases.
